requirements.txt
Adafruit-Blinka==8.46.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-neopixel-spi==1.0.9
adafruit-circuitpython-pixelbuf==2.0.4
adafruit-circuitpython-requests==4.1.4
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.72.1
Adafruit-PureIO==1.1.11
asttokens==2.4.1
binho-host-adapter==0.1.6
board==1.0
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
einops==0.8.0
executing==2.0.1
filelock==3.15.4
fsspec==2024.6.1
huggingface_hub==0.23.4
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyftdi==0.55.4
Pygments==2.18.0
pyserial==3.5
python-dateutil==2.9.0.post0
pyusb==1.2.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
safetensors==0.4.3
six==1.16.0
stack-data==0.6.3
sympy==1.12.1
sysv-ipc==1.1.0
timm==1.0.7
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
